+++ clang/lib/Basic/Targets/AArch64.cpp
@@ -1445,6 +1445,7 @@ AArch64TargetInfo::checkCallingConvention(CallingConv CC) const {
   case CC_PreserveMost:
   case CC_PreserveAll:
   case CC_PreserveNone:
+  case CC_CoroHandleFn:
   case CC_DeviceKernel:
   case CC_AArch64VectorCall:
   case CC_AArch64SVEPCS:
@@ -1778,6 +1779,7 @@ WindowsARM64TargetInfo::checkCallingConvention(CallingConv CC) const {
   case CC_PreserveMost:
   case CC_PreserveAll:
   case CC_PreserveNone:
+  case CC_CoroHandleFn:
   case CC_Swift:
   case CC_SwiftAsync:
   case CC_Win64: